+++ browse/src/main/kotlin/com/alfresco/content/browse/tasks/attachments/AttachedFilesFragment.kt
@@ -5,10 +5,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.appcompat.app.AlertDialog
-import androidx.core.content.res.ResourcesCompat
 import androidx.core.view.isVisible
-import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import com.airbnb.epoxy.AsyncEpoxyController
 import com.airbnb.mvrx.Loading
@@ -19,29 +16,25 @@ import com.alfresco.content.actions.ActionOpenWith
 import com.alfresco.content.browse.R
 import com.alfresco.content.browse.databinding.FragmentAttachedFilesBinding
 import com.alfresco.content.browse.preview.LocalPreviewActivity
+import com.alfresco.content.browse.tasks.BaseDetailFragment
 import com.alfresco.content.browse.tasks.detail.TaskDetailViewModel
-import com.alfresco.content.browse.tasks.detail.execute
 import com.alfresco.content.data.AnalyticsManager
 import com.alfresco.content.data.ContentEntry
 import com.alfresco.content.data.Entry
-import com.alfresco.content.data.EventName
 import com.alfresco.content.data.PageView
 import com.alfresco.content.data.ParentEntry
 import com.alfresco.content.listview.EntryListener
 import com.alfresco.content.mimetype.MimeType
 import com.alfresco.content.simpleController
 import com.alfresco.ui.getDrawableForAttribute
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import java.lang.ref.WeakReference
 
 /**
  * Marked as AttachedFilesFragment class
  */
-class AttachedFilesFragment : Fragment(), MavericksView, EntryListener {
+class AttachedFilesFragment : BaseDetailFragment(), MavericksView, EntryListener {
 
     val viewModel: TaskDetailViewModel by activityViewModel()
     private lateinit var binding: FragmentAttachedFilesBinding
-    private var deleteContentDialog = WeakReference<AlertDialog>(null)
     private val epoxyController: AsyncEpoxyController by lazy { epoxyController() }
 
     override fun onCreateView(
@@ -82,11 +75,15 @@ class AttachedFilesFragment : Fragment(), MavericksView, EntryListener {
         viewModel.setListener(this)
     }
 
+    override fun onConfirmDelete(contentId: String) {
+        viewModel.deleteAttachment(contentId)
+    }
+
     override fun invalidate() = withState(viewModel) { state ->
 
         binding.loading.isVisible = state.requestDeleteContent is Loading
 
-            if (state.requestContents.complete) {
+        if (state.requestContents.complete) {
             binding.refreshLayout.isRefreshing = false
         }
 
@@ -116,28 +113,6 @@ class AttachedFilesFragment : Fragment(), MavericksView, EntryListener {
 
     private fun onItemClicked(contentEntry: ContentEntry) {
         viewModel.execute(ActionOpenWith(Entry.convertContentEntryToEntry(contentEntry, MimeType.isDocFile(contentEntry.mimeType))))
-    }
-
-    private fun deleteContentPrompt(contentEntry: ContentEntry) {
-        AnalyticsManager().taskEvent(EventName.DeleteTaskAttachment)
-        val oldDialog = deleteContentDialog.get()
-        if (oldDialog != null && oldDialog.isShowing) return
-        val dialog = MaterialAlertDialogBuilder(requireContext())
-            .setTitle(getString(R.string.dialog_title_delete_content))
-            .setMessage(contentEntry.name)
-            .setIcon(
-                ResourcesCompat.getDrawable(
-                    requireContext().resources,
-                    MimeType.with(contentEntry.mimeType).icon,
-                    requireContext().theme
-                )
-            )
-            .setNegativeButton(getString(R.string.dialog_negative_button_task), null)
-            .setPositiveButton(getString(R.string.dialog_positive_button_task)) { _, _ ->
-                viewModel.deleteAttachment(contentEntry.id.toString())
-            }
-            .show()
-        deleteContentDialog = WeakReference(dialog)
     }
 
     override fun onEntryCreated(entry: ParentEntry) {
